requirements.txt
pandas
numpy
scikit-learn
tensorflow
fastapi
uvicorn
pyyaml
joblib
pydantic